Rename server logger
