+++ crates/cubecl-runtime/src/lib.rs
@@ -35,8 +35,8 @@ pub use base::*;
 pub use cubecl_common::benchmark;
 
 pub use feature_set::*;
-/// Debugging utilities.
-pub mod debug;
+/// Logging utilities to be used by a compute server.
+pub mod logging;
 
 /// TMA-related runtime types
 pub mod tma;
