+++ crates/cubecl-runtime/src/logging/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "std")]
+mod profiling;
+#[cfg(not(feature = "std"))]
+mod profiling {}
+pub use profiling::*;
+
+mod server;
+
+pub use server::*;
