+++ crates/cubecl-wgpu/src/runtime.rs
@@ -12,7 +12,7 @@ use cubecl_runtime::{
     ComputeRuntime, TimeMeasurement,
     channel::MutexComputeChannel,
     client::ComputeClient,
-    debug::{DebugLogger, ProfileLevel},
+    logging::{ProfileLevel, SeverLogger},
 };
 use cubecl_runtime::{DeviceProperties, memory_management::HardwareProperties};
 use cubecl_runtime::{memory_management::MemoryDeviceProperties, storage::ComputeStorage};
@@ -302,8 +302,8 @@ async fn request_adapter(
     device: &WgpuDevice,
     backend: wgpu::Backend,
 ) -> (wgpu::Instance, wgpu::Adapter) {
-    let debug = DebugLogger::default();
-    let instance_flags = match (debug.profile_level(), debug.is_activated()) {
+    let debug = SeverLogger::default();
+    let instance_flags = match (debug.profile_level(), debug.compilation_activated()) {
         (Some(ProfileLevel::Full), _) => InstanceFlags::advanced_debugging(),
         (_, true) => InstanceFlags::debugging(),
         (_, false) => InstanceFlags::default(),
